ficed page deletion

+++ src/main/java/com/example/marketour/services/TourPagesService.java
@@ -2,22 +2,22 @@ package com.example.marketour.services;
 
 import com.example.marketour.model.entities.*;
 import com.example.marketour.repositories.TourPageRepository;
+import com.example.marketour.repositories.TourRepository;
 import org.springframework.stereotype.Service;
 
-import java.util.Comparator;
 import java.util.List;
-import java.util.Map;
 import java.util.Objects;
-import java.util.logging.Level;
-import java.util.logging.Logger;
 import java.util.stream.Collectors;
 
 @Service
 public class TourPagesService {
     private final TourPageRepository tourPageRepository;
+    private final TourRepository tourRepository;
 
-    public TourPagesService(TourPageRepository tourPageRepository) {
+    public TourPagesService(TourPageRepository tourPageRepository,
+                            TourRepository tourRepository) {
         this.tourPageRepository = tourPageRepository;
+        this.tourRepository = tourRepository;
     }
 
     public List<TourPage> getAllTourPages(Long tourId) {
@@ -32,7 +32,6 @@ public class TourPagesService {
         final var old = tourPageRepository.findAll().stream()
                 .filter(tourPage -> Objects.equals(tourPage.getTour().getTourId(), tourId))
                 .collect(Collectors.toList());
-        Logger.getGlobal().log(Level.INFO, old.stream().sorted(Comparator.comparing(TourPage::getTourPageId)).map(tourPage -> Map.entry(tourPage.getTourPageId(), tourPage.getPage())).collect(Collectors.toMap(Map.Entry::getKey, Map.Entry::getValue)).toString());
         final var newPages = newIds.stream().map(aLong -> {
             final var occurrence = old.stream().filter(tourPage -> Objects.equals(tourPage.getTourPageId(), aLong)).findFirst().orElse(null);
             if (occurrence != null) {
@@ -41,7 +40,6 @@ public class TourPagesService {
             return occurrence;
         }).collect(Collectors.toList());
 
-        Logger.getGlobal().log(Level.INFO, newPages.stream().sorted(Comparator.comparing(TourPage::getTourPageId)).map(tourPage -> Map.entry(tourPage.getTourPageId(), tourPage.getPage())).collect(Collectors.toMap(Map.Entry::getKey, Map.Entry::getValue)).toString());
         tourPageRepository.saveAll(newPages);
     }
 
@@ -76,6 +74,12 @@ public class TourPagesService {
     }
 
     public void removeTourPageById(Long tourPageId) {
-        tourPageRepository.deleteById(tourPageId);
+        final var tourPage = tourPageRepository.findById(tourPageId).orElse(null);
+        if (tourPage != null) {
+            tourPage.getTour().getTourPages().remove(tourPage);
+            tourRepository.save(tourPage.getTour());
+            tourPageRepository.deleteById(tourPageId);
+        }
+
     }
 }
